Explicitly make global variables global
Using Homebrew Python 2.7.10 on OS X, unless these variables are marked as global in `main` they become implicitly local since they are assigned. This prevents the hook from working because the implicitly local variables shadow their implicitly global counterparts in the rest of the code, and the global variables are never assigned.